PotentioMeter clean up
